legacy/alphaflow/requirements.txt
# ====================================
# AlphaFlow - Requirements
# ====================================

# Native macOS App
PyQt6>=6.5.0
PyQt6-WebEngine>=6.5.0
PyQt6-Charts>=6.5.0

# Web App Framework
streamlit>=1.29.0
plotly>=5.18.0

# Core Dependencies (Required)
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
websockets>=11.0
scikit-learn>=1.3.0
scipy>=1.11.0
pytz>=2023.3

# Trading APIs
alpaca-trade-api>=3.0.0

# Technical Analysis (Optional - will use fallback if not available)
# TA-Lib>=0.4.28

# Deep Learning
torch>=2.0.0
transformers>=4.30.0

# Sentiment Analysis
nltk>=3.8.0
textblob>=0.17.1
newsapi-python>=0.2.7
tweepy>=4.14.0

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-qt>=4.2.0
